requirements.txt
# Core dependencies
orjson>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Web automation
playwright>=1.40.0
tenacity>=8.2.0

# HTTP requests (for API calls)
requests>=2.31.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Image processing (for screenshots)
pillow>=10.0.0

# Data processing (optional)
pandas>=2.0.0
numpy>=1.24.0